Introduce a client-server model
With the introduction of 47fdd45a (Default guest memory to 80% of total
RAM on host), we're targeting a use case in which we expect to be
running a single, large VM.

To accommodate this use case, we're switching here to a client/server
model. The first time krun is launched, in addition to running the
command, it starts a small server in the VM listening on a TCP port. On
the host side, it also creates a lock file, writes the listening port
number to it, and calls flock() on it to prevent other instances from
running in parallel.

When it's executed again, krun detects there's another instance running,
and contacts the server in the VM to request the launch of the command.
The stdout and stderr of the command are redirected to
$XDG_RUNTIME_DIR/krun-$COMMAND-$TIMESTAMP.[stdout|stderr]

Co-authored-by: Teoh Han Hui <teohhanhui@gmail.com>
Signed-off-by: Sergio Lopez <slp@sinrega.org>