Autoconfigure network without DHCP
Passt includes a DHCP server to allow guests to easily configure the
network without being aware passt is on the other side. But we _are_
aware of that, so we can take advantage of it.

Instead of using DHCP, read the configuration output from passt,
marshall it into some environment variables, pass it to the guest, and
have krun-guest read it and apply it using rtnetlink.

By doing that, we reduce the startup time in half, from...

$ time krun /bin/false
(...)
real    0m4,301s

... to ...

$ time krun /bin/false
(...)
real    0m1,966s

In addition of reducing the boot time, this potentially will prevent
some of the dhcp issues we've seen in the past.

Signed-off-by: Sergio Lopez <slp@redhat.com>