Pass args and envs using a config file
When using "krun_set_exec", the arguments and environment variables are
passed to the guest using the kernel command line. In addition to being
limited in length, this is also only compatible with ASCII strings.

libkrun is also capable or reading both arguments and environment
variables from a JSON config file, which doesn't suffer from any of
those limitations, so switch to using this approach instead.

Depends on https://github.com/containers/libkrun/pull/220

Signed-off-by: Sergio Lopez <slp@redhat.com>