x11: Introduce the -x command line option to enable X11 forwarding
Instead of enabling X11 forwarding unconditionally but hiding it behind
the HOST_DISPLAY variable, enable it if the --direct-x11 flag is used.
In that case, directly set it as the DISPLAY environment variable.

This makes it easier for users to enable the feature and use it
correctly. Eventually, we can flip the logic of the flag to make direct
X11 the default.

Keep the server at :1 to keep churn down, and because it seems like a
generally helpful distinction, since the server is usually :0 on the
host.

Signed-off-by: Asahi Lina <lina@asahilina.net>