mount: Add support for FEX merged rootfs mode
In this mode, the FEX rootfs passed to FEX is already overlaid on top of
the real root filesystem, so FEX directs all guest accesses to it
instead of doing its own overlay logic. This, together with a bunch of
fixes on the FEX side, fixes Wine.

Opt-in for now, since this actively *breaks* things with the current
FEX. May become the default in the future once all that is sorted out.

Signed-off-by: Asahi Lina <lina@asahilina.net>