Use the same server instance for privileged launches too
We can keep the root uid in our saved set user id, and copy it to ruid/euid
after fork, if launching as root is requested. Writing to drop_caches becomes
a bit tricker when running unprivileged, but we can acquire the privileges in
a forked process for it to work and not affect other launches.

Signed-off-by: Sasha Finkelstein <fnkl.kernel@gmail.com>