Support using Box64 for x86 emulation
Add the "emu" command line argument to let users choose the emulator to
be used for running x86 binaries. Add support for installing Box64 in
binfmt_misc. If the argument is not present in the command line, try to
use FEX first and if it can't be set up, try again with Box64.

Supersedes: #140
Co-authored-by: ptitSeb <sebastien.chev@gmail.com>
Co-authored-by: Alex Arnold <m1m1k4tz@protonmail.com>
Signed-off-by: Sergio Lopez <slp@redhat.com>