Add a flag to run a custom init instead of muvm-guest
Allow the user to replace muvm-guest with a custom init process. With
updated libkrun, it will even run as PID 1, making it possible to run
systemd.

Of course, this is not the suggested way to use muvm, but some use cases
necesitate the use of systemd. The next few commits will facilitate
running individual parts of muvm-guest as separate processes under a
custom service manager.

Signed-off-by: Val Packett <val@invisiblethingslab.com>