muvm-guest: pwbridge: actually read fd indexes in the Transport msg
The message payload contains indexes for the attached file descriptors.
We must use them instead of relying on the order, which was never guaranteed,
and *actually* differs in practice between architectures in compilers
because current PipeWire code calls the fd-attaching function inside of
function call arguments, which *don't* have a defined evaluation order
in the C language standard.

Signed-off-by: Val Packett <val@invisiblethingslab.com>